requirement.txt
absl-py==1.4.0
appdirs==1.4.4
astunparse==1.6.3
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.0.1
click==8.1.6
cycler==0.11.0
docker-pycreds==0.4.0
flatbuffers==23.3.3
fonttools==4.38.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.5
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.3.5
pathtools==0.1.2
Pillow==9.5.0
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-version==0.0.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.29.2
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
torch==1.11.0+cu113
torch-cluster==1.6.0
torch-geometric==2.2.0
torch-scatter==2.1.0
torch-sparse==0.6.16
torch-spline-conv==1.2.1
torchsummary==1.5.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
wandb==0.15.8
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
